updates
focal -> jammy
alpine3.15 to alpine 3.17
libimagequant 4.0.1 to to 4.1.1
golang 1.16 to 1.19
