+++ docker/alpine/Dockerfile
@@ -6,13 +6,13 @@ RUN /root/enable-world-repos.sh
 RUN apk update && \
     apk upgrade && \
     apk add rust cargo cargo-c bash curl bind-tools sudo build-base git musl-dev libressl-dev && \
-    git clone https://github.com/ImageOptim/libimagequant.git --branch 4.0.1
+    git clone https://github.com/ImageOptim/libimagequant.git --branch 4.1.1
 
 RUN /bin/sh -c 'cd libimagequant/imagequant-sys && \
     cargo cinstall --destdir=. '
 
 # compile
-FROM golang:1.18.8-alpine3.17 as builder
+FROM golang:1.19-alpine3.17 as builder
 LABEL maintainer="Christian Juerges, christian.juerges@tx.group"
 LABEL stage=builder
 
@@ -32,7 +32,7 @@ COPY . .
 RUN /bin/sh -c 'make test'
 RUN /bin/sh -c 'make build'
 
-# build image, based on alpine:3.15
+# build image, based on alpine
 FROM alpine:3.17
 LABEL maintainer="Christian Juerges, christian.juerges@tx.group"
