+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/DND-IT/go-imagequant
 
-go 1.16
+go 1.19
 
-require golang.org/x/image v0.0.0-20220321031419-a8550c1d254a
+require golang.org/x/image v0.6.0
