--- lib/ubuntu/20.04/lib/libimagequant.so
@@ -1,1 +1,0 @@
-libimagequant.so.0.0.0
