--- lib/ubuntu/20.04/lib/pkgconfig/imagequant.pc
@@ -1,11 +1,0 @@
-prefix=/usr/local
-exec_prefix=${prefix}
-libdir=${exec_prefix}/lib
-includedir=${prefix}/include
-
-Name: imagequant
-Description: Convert 24/32-bit images to 8-bit palette with alpha channel.
-Version: 4.0.0
-Libs: -L${libdir} -limagequant
-Cflags: -I${includedir}
-Libs.private:  -lgcc_s -lutil -lrt -lpthread -lm -ldl -lc
