+++ lib/ubuntu/22.04/lib/libimagequant.so
@@ -1,0 +1,1 @@
+libimagequant.so.0.0.0
